% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{docker_quick_start}
\alias{docker_quick_start}
\title{Quick running docker container}
\usage{
docker_quick_start(image, name = NULL, theme_succession = FALSE, ...)
}
\arguments{
\item{image}{docker image name (character)}

\item{name}{docker container name (character). By default the project's name will be used.}

\item{theme_succession}{a logical value indicating whether employ local RStudio theme & setting.}

\item{...}{other parameter set to docker::docker$clientcontainers$run object}
}
\description{
Quick running docker container
}
\examples{
\dontrun{
rss_instance <- docker_quick_start(image = "uribo/tiny_rocker_minimum",
remove  = TRUE,
detach  = TRUE)
rss_instance$start()
rss_instance$stop()
}
}
